rename core filen
